fix(tests): address codex review P1/P2 — tracking ref and T-1 consolidation
P1: add 'tracked: #1130' to skipif reason in test_health_scorer_helpers_encoding.py
    so the Windows-only real-locale check satisfies the skip-tracking governance rule.

P2: consolidate encoding regression tests into existing test_r37au_ap001_constructor_kwarg_suppression.py
    per T-1 rule (tests for an existing MCP utility belong in the existing file).
    Remove test_anti_patterns_encoding.py (was a new file, violates T-1).
